feat: complete env validation with proper startup sequence
