+++ backend/app.js
@@ -3,17 +3,8 @@ import morgan from 'morgan';
 import process from 'node:process';
 import path from 'node:path';
 import apiRouter from './controllers/index.js';
-import { validateEnv } from './config/config.js'; // Import the validator
 
 const app = express();
-
-// Validate environment before proceeding
-const missingVars = validateEnv();
-if (missingVars.length > 0) {
-  console.error('❌ Server cannot start:');
-  missingVars.forEach((varName) => console.error(`   ${varName} is required`));
-  process.exit(1);
-}
 
 app.use(express.json());
